feat/ Add support for GPT-4.1 family (#834)
Adds support for the following models:
- `gpt-4.1` (Latest)
- `gpt-4.1-mini` (Latest)
- `gpt-4.1-nano` (Latest)
- `gpt-4.1-2025-04-14` (Pinned)
- `gpt-4.1-mini-2025-04-14` (Pinned)
- `gpt-4.1-nano-2025-04-14` (Pinned)

Changes the `sort_order` property from `int` to `float` so that sorting
can be done by family (int part for model family sorting and decimal
part for sorting within a model family).

+++ pingpong/schemas.py
@@ -1018,7 +1018,7 @@ class AssistantModel(BaseModel):
 
 class AssistantModelDict(TypedDict):
     name: str
-    sort_order: int
+    sort_order: float
     is_latest: bool
     is_new: bool
     highlight: bool
